+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## Unreleased
+* minification
 * XHR constructor parameter support for Gecko-powered browsers
 
 ## [0.9.0-alpha.4–9](https://github.com/Mouvedia/cb-fetch/compare/b15a26f...d5c09ea)