+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## Unreleased
+* XDR
 * timeout callback
 * automatic document parsing
 * XSL pattern query language option for Trident-based browsers